refactor folder structure, extract components to their own files
