+++ frontend/src/lib/components/AppEditorLink/utils.js
@@ -1,0 +1,4 @@
+export function appEditorUrl (actionId, appUrl) {
+    return '/api/user/redirect_to_site/' + (actionId ? '?actionId=' + actionId : '') + (appUrl ? `${actionId ? '&' : '?'}appUrl=${encodeURIComponent(appUrl)}` : '')
+}
+export const defaultUrl = 'https://'
