+++ frontend/src/lib/components/Dropdown.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+
 export class Dropdown extends Component {
     constructor(props) {
         super(props);
