+++ frontend/src/lib/components/LoadingOverlay.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export function LoadingOverlay () {
+  return (
+      <div className='loading-overlay' style={{ paddingTop: '14rem' }}>
+          <div />
+          <br />
+          (This might take a while)
+      </div>
+  )
+}
