+++ frontend/src/lib/components/Modal.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 
-export default class Modal extends Component {
+export class Modal extends Component {
     constructor(props) {
         super(props)
