+++ frontend/src/scenes/trends/BreakdownFilter.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import Select from 'react-select'
 import { selectStyle } from '../../lib/utils'
 
-export default class BreakdownFilter extends Component {
+export class BreakdownFilter extends Component {
   render() {
     return <div style={{width: 200, display: 'inline-block'}}>
       <Select
